Moving reusable files into their own utils folder
