+++ client/package.json
@@ -18,8 +18,6 @@
     "@google-cloud/functions-framework": "^1.7.1",
     "@google-cloud/storage": "^5.8.3",
     "discord.js": "^12.5.3",
-    "dotenv": "^8.2.0",
-    "express": "^4.17.1",
     "google-auth-library": "^7.0.4",
     "slash-commands": "^1.4.0",
     "tweetnacl": "^1.0.3",
