+++ client/src/delete.test.ts
@@ -1,9 +1,9 @@
-import { createAuthenticatedClient } from "./authentication";
-import { handleRequest } from "./requestHandler";
+import { createAuthenticatedClient } from "./utils/authentication";
+import { handleRequest } from "./utils/requestHandler";
 import { deleteChannel } from "./delete";
 
-jest.mock("./authentication");
-jest.mock("./requestHandler");
+jest.mock("./utils/authentication");
+jest.mock("./utils/requestHandler");
 
 const mockCreateAuthenticatedClient = createAuthenticatedClient as jest.MockedFunction<
   typeof createAuthenticatedClient
