+++ client/src/utils/authentication.test.ts
