+++ client/src/utils/authentication.ts
